Added option for doing a dry-run
This allows one to see what webgen would do but without writing
anything.

+++ lib/webgen/website.rb
@@ -178,6 +178,7 @@ module Webgen
 
     # Save the +cache+.
     def save_cache
+      return if config['website.dry_run']
       cache_data = [@cache.dump, Webgen::VERSION]
       if config['website.cache'].first == :file
         File.open(cache_file, 'wb') {|f| Marshal.dump(cache_data, f)}
